A few bug fixes to implement the dpg11 cwodmr driver, including a bug where the plot_waveforms was not showing the right user_defined channels, and the binary functions not properly convering things to integer arrays, as should be the case.
Things are looking good for now, I will test the outputs tomorrow, and make sure that the wavefiles being in folders is not a huge issue. Things are looking really good as of now though.

+++ dpg11_pylib/waveforms/waveforms.py
@@ -140,9 +140,9 @@ def pad_waveform(waveform: np.ndarray or list,
     
     # Pad the waveform
     if pad_side == 'right':
-        waveform = np.pad(waveform, (pad_value, num_samples_to_pad), constant_values=pad_value)
+        waveform = np.pad(waveform, (pad_value, num_samples_to_pad), constant_values=pad_value).astype(int)
     elif pad_side == 'left':
-        waveform = np.pad(waveform, (num_samples_to_pad, pad_value), constant_values=pad_value)
+        waveform = np.pad(waveform, (num_samples_to_pad, pad_value), constant_values=pad_value).astype(int)
     else:
         raise ValueError('pad_side must be either "right" or "left"')
